changed vre to cre merge and close #11
modules are always constants. Declaring them with var only makes sense in node.js 0.12 and lesser where const isn't available.
This is described as "A collection of ES6 snippets for faster JavaScript development" in the main readme so I think it is okay to promote the best practices incompatible with 0.12. I think most people using this collection of snippets are already on newer node.js than 0.12.